added functions in modules
